Modified lexical_analysis to read PRINT statements
In this iteration, I attempted to read the PRINT statements without regex. This required specific code to read PRINT statements when first analysing the tokens. This implementation has bugs and the program gets caught in regex issues when operators are lead and followed by spaces. Possible solutions are updating the regex or rethinking the approach to PRINT statements