Implemented division and cleaned up variable assignment
An oversight that was corrected was if you had a defined variable, say "NUM = 10", then you tried to create a new variable with an operator, the program would throw an error since the variable wasn't already defined. So before, "VAR += NUM" wouldn't work since 'VAR' was undefined. Now, the default value is set to 0 so "VAR += NUM" is essentially 0 + NUM and doesn't throw an error.