Continued implementation of FOR LOOP
Currently, all other files work except code4.zpm, where the print statement repeats. This might be due to debugging statements I have, but in this iteration, FOR LOOP logic is close to working. The iteration count and loop body are seperated, and are to be parsed individually. The error arises when the statements to be parsed are interpreted incorrectly by the lexical analysis.